refactor ctxutil to httputil
