+++ internal/httputil/context.go
@@ -1,4 +1,4 @@
-package ctxutil
+package httputil
 
 import (
 	"context"
